added mosaic generation

+++ src/main/java/com/plywacz/mosaic/services/ImageService.java
@@ -4,8 +4,6 @@ Author: BeGieU
 Date: 15.12.2019
 */
 
-import java.awt.image.BufferedImage;
-
 public interface ImageService {
-    byte[] generateMosaic(Long random, String resolution, String images);
+    byte[] generateMosaic(Integer random, String resolution, String images);
 }
